+++ adekz/README.md
@@ -1,33 +1,16 @@
 #Study of NextJS and documentation
-<br>
-First: install NodeJS 
-<br>
-Second: npx create-next-app 
-<br>
+First: install NodeJS
+Second: npx create-next-app
 Third: npm run dev
-<br>
-#I'm using gitflow to organize the project
-https://www.atlassian.com/br/git/tutorials/comparing-workflows/gitflow-workflow
-<br>
-create folder SRC 
-<br>
+#I'm using gitflow to organize the project https://www.atlassian.com/br/git/tutorials/comparing-workflows/gitflow-workflow
+create folder SRC
 Next js magic, component import routes don't need to be organized / hahaha
-<br>
 create folder components
-<br>
 what's is next JS ? nextjs is the framewor version of react
-<br>
 #Organization folder
-<br>
 #styles
-<br>
 global.css / Global styles
-<br>
 Home.module.css / Style of modules and files
-<br>
 The public folder is for images and design files.
-<br>
 NextJs is a fullstack framework
-<br>
 In the next JS the route settings are directed to the page folder, you don't need to configure that beautiful
-<br>
