+++ adekz/package.json
@@ -17,5 +17,4 @@
     "eslint": "7.32.0",
     "eslint-config-next": "11.1.2"
   }
-  
 }
