--- adekz/src/pages/bacabal.jsx
@@ -1,6 +1,0 @@
-export default function bacabal() {
-    return (
-     "Bacabal"
-    )
-  }
-
