--- adekz/src/pages/index.jsx
@@ -1,11 +1,0 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import api from './bacabal'
-
-export default function Home() {
-  return (
-   <div>
-     Não sei o que fazer hahahahaha
-   </div>
-  )
-}
